+++ src/main/java/io/jenkins/plugins/analysis/core/util/ConsoleLogReaderFactory.java
@@ -2,6 +2,10 @@ package io.jenkins.plugins.analysis.core.util;
 
 import java.io.IOException;
 import java.io.Reader;
+import java.util.function.Function;
+import java.util.regex.Pattern;
+
+import org.apache.commons.lang3.StringUtils;
 
 import com.google.errorprone.annotations.MustBeClosed;
 
@@ -19,6 +23,11 @@ import hudson.model.Run;
 public class ConsoleLogReaderFactory extends ReaderFactory {
     private final Run<?, ?> run;
 
+    private static final Pattern TIME_STAMPER_PREFIX
+            = Pattern.compile("\\[\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z] ");
+    private static final Function<String, String> REMOVE_TIME_STAMPER_PREFIX
+            = string -> TIME_STAMPER_PREFIX.matcher(string).replaceFirst(StringUtils.EMPTY);
+
     /**
      * Creates a new {@link ConsoleLogReaderFactory}.
      *
@@ -26,7 +35,7 @@ public class ConsoleLogReaderFactory extends ReaderFactory {
      *         the run that provides the console log
      */
     public ConsoleLogReaderFactory(final Run<?, ?> run) {
-        super(run.getCharset(), ConsoleNote::removeNotes);
+        super(run.getCharset(), REMOVE_TIME_STAMPER_PREFIX.compose(ConsoleNote::removeNotes));
 
         this.run = run;
     }